+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 <h3>General</h3>
 * Added better exception handling
 
+<h3>ORM</h3>
+* Made it easier to add repositories to `OrmBootstrapper`
+
 <h2>v1.0.1 (2017-01-02)</h2>
 
 <h3>General</h3>